Fix : 오디오 세팅 게임매니저에서 전역적으로 관리
