+++ Assets/Resources/PopUp/SettingPopUp.prefab
@@ -182,9 +182,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 29606a5fdab77d8419f7217bab776bf6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  audioMixer: {fileID: 24100000, guid: 4e7e7b7a716e99840b9ecb1c59e2ff2c, type: 2}
-  bgmSlider: {fileID: 4434372993938879392}
-  sfxSlider: {fileID: 2407963844104482886}
+  _bgmSlider: {fileID: 4434372993938879392}
+  _sfxSlider: {fileID: 2407963844104482886}
 --- !u!1 &1434506389571003025
 GameObject:
   m_ObjectHideFlags: 0
